requirements.txt
google-genai>=1.24.0
python-dotenv>=1.1.0
pandas>=2.2
openpyxl>=3.1.5
psycopg2>=2.9.10